Change handleMouseMoved argument type from float to double
Also rename functions to camelCase.

+++ gui/include/graphics/interface.hpp
@@ -26,7 +26,7 @@ class Interface : public RenderTarget, public InputHandler {
     bool handleScroll(double offset) override;
     bool handleKey(int key, bool down, int mods) override;
     bool handleChar(unsigned int c) override;
-    bool handleMouseMoved(float x, float y) override;
+    bool handleMouseMoved(double x, double y) override;
 
     int priority() const override { return 1; }
