+++ gui/include/graphics/scene_camera_3d.hpp
@@ -45,7 +45,7 @@ class SceneCamera3d : public SceneCamera {
 
     bool handleMouseButton(int button, bool down) override;
     bool handleScroll(double offset) override;
-    bool handleMouseMoved(float x, float y) override;
+    bool handleMouseMoved(double x, double y) override;
     bool handleKey(int key, bool down, int mods) override;
 
     void tick(float time_elapsed) override;
@@ -67,8 +67,8 @@ class SceneCamera3d : public SceneCamera {
     float angle_ = 0.0f;
     float scale_ = 0.5f;
 
-    float prev_x_ = -1.1f;
-    float prev_y_ = -1.1f;
+    double prev_x_ = -1.1;
+    double prev_y_ = -1.1;
     glm::vec2 delta_;
 
     bool dragging_ = false;
